feat(go/lint): upgrade go to 1.16 & add golangci-lint
