+++ .gitignore
@@ -24,3 +24,4 @@ _testmain.go
 tags
 .envrc
 vendor
+.coverage.cov
