+++ Makefile
@@ -5,7 +5,24 @@ PACKAGES ?= $(shell $(GO) list ./...)
 VETPACKAGES ?= $(shell $(GO) list ./... | grep -v /examples/)
 GOFILES := $(shell find . -name "*.go")
 
+.PHONY: init
+init:
+	@if [ $(GO_VERSION) -gt 15 ]; then \
+		$(GO) install github.com/golangci/golangci-lint/cmd/golangci-lint@v1.43.0; \
+	elif [ $(GO_VERSION) -lt 16 ]; then \
+		$(GO) get -u github.com/golangci/golangci-lint/cmd/golangci-lint@v1.40.1; \
+	fi
+
 .PHONY: dep
 dep:
-	go mod tidy
-	go mod vendor
+	$(GO) mod tidy
+	$(GO) mod vendor
+
+.PHONY: lint
+lint:
+	export GOFLAGS=-mod=vendor
+	golangci-lint run
+
+.PHONY: test
+test:
+	$(GO) test -mod=vendor . -covermode=count -coverprofile .coverage.cov
