+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/wklken/gorequest
 
-go 1.13
+go 1.16
 
 require (
 	github.com/elazarl/goproxy v0.0.0-20211114080932-d06c3be7c11b
